Fix for https://github.com/Sybit-Education/airtable.java/issues/43
Previously a warning was logged if Table.setProperty was called with a keyname that was not a valid java identifier (meaning it has illegal characters preventing it from being mapped directly to a class field).  Additionally, if Table.setProperty was called with a keyname that did not match a field for type T, setProperty failed with a warning only.

Changed Table.setProperty to (1) more comprehensively detect illegal characters; (2) provide more informative log message on error condition; and (3) now a Runtime exception is thrown if the keyname cannot be matched to a field on type T.

Added tests in TableTest.java for the setProperty method.